+++ CHANGES.md
@@ -4,6 +4,7 @@
 - The Architect Table in the Garrison will now automatically designate a plot for the Barracks, provided that the player's Garrison is at level 1 and they are currently on the "Build Your Barracks" quest.
 - The Mission Table in the Garrison will now have its first mission automated, provided the player has yet to send their first mission, the player's Garrison is level 1, and the player is currently on the "Delegating on Draenor" (Alliance) or "Mission Probable" (Horde) quest.
 ### Changed or Updated
+- Users can now shift-click on the micro menu button to access the Talent Importer utility. The tooltip has been updated to indicate this new functionality.
 ### Deprecated
 ### Removed
 ### Fixed